requirements.txt
Django==1.11.13
# admin
django-suit==0.2.25
django-ckeditor==5.3.0
Pillow==5.2.0
# Heroku dependencies
dj-database-url==0.5.0
gunicorn==19.7.1
django-heroku==0.3.1
# S3 Dependencies
boto3==1.9.7
django-storages==1.7.1


